[nl] Fix the text on the potion contest pamphlet
The text on the pamphlet was not shown correctly, since the font was
changed to another one since I translated this. This commit fixes that.

Furthermore, when I translated this, I interpreted 'Azarday, 3 Pinkmoon'
as 'Azarday' being the weekday and '3 Pinkmoon' being the date (as in
'Tuesday, 3 January'). Now I read on
https://github.com/Deevad/peppercarrot/wiki/Time-system
that Pinkmoon actually indicates time, so I changed the Dutch to reflect
this.